refactor: new game structure
Changed how the game internally works by making a proper main scene and isolating the multiplayer controller into it's own thing. Refactored the rest of the project to match this with a ton of fixes around the board.

Continued work on the deck and card system.

Chose to make a more "centralized" approach to everything, where nodes there will only be one of (singleton per game instance) will now live in the game main scene and not be isolated like the hud was before.